Add a timeout to balloon. Pass this to contents
Great plugin thanks. Some suggested changes that I have found useful. 

I have changed the call to a contents function so that is sets this to be the target. This makes interactive contents easier.
I have also added a timeout that can be set which will hide the balloon after a set number of milliseconds even if mouse is still over.